requirements.txt
streamlit
pandas
paho-mqtt
pydeck